aufs: inode op, add, link
Implement i_op->link().
As aufs supports 'pseudo-link', aufs_link() can make it without
copying-up. In the case of aufs_link() has to copy-up, the name of the
target file is used as-is, and it is pseudo-linked. In other words,
calling link(2) after the copy-up is unnecessary.

See also struct.txt in previous commit.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>