aufs: fhsm (file-based hierarchical storage management)
This feature automatically handles MVDOWN in other commits.
In user-space, a daemon monitors the free space of the branch and issues
MVDOWN ioctl automatically when necessary. The main role is in
user-space and several options are implemented.
For a branch to join the FHSM circle, a new attribute 'fhsm' should be
specified.

See also the document in this commit.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>